+++ Sources/Sentry/SentryClient.m
@@ -81,10 +81,10 @@ NSString *const DropSessionLogMessage = @"Session has no release name. Won't sen
 
         NSError *error = nil;
 
-        self.fileManager =
-            [[SentryFileManager alloc] initWithDsn:self.options.parsedDsn
-                            andCurrentDateProvider:[[SentryDefaultCurrentDateProvider alloc] init]
-                                  didFailWithError:&error];
+        self.fileManager = [[SentryFileManager alloc]
+                   initWithOptions:self.options
+            andCurrentDateProvider:[[SentryDefaultCurrentDateProvider alloc] init]
+                             error:&error];
         if (nil != error) {
             [SentryLog logWithMessage:error.localizedDescription andLevel:kSentryLevelError];
             return nil;