+++ Sources/Sentry/SentryCrashIntegration.m
@@ -7,7 +7,7 @@
 #import "SentryHook.h"
 #import "SentryHub.h"
 #import "SentryOutOfMemoryLogic.h"
-#import "SentrySDK.h"
+#import "SentrySDK+Private.h"
 #import "SentryScope+Private.h"
 #import "SentrySessionCrashedHandler.h"