+++ Tests/SentryTests/SentrySDK+Tests.h
@@ -1,0 +1,15 @@
+#import <Sentry/Sentry.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface SentrySDK (Tests)
+
++ (void)setCurrentHub:(SentryHub *)hub;
+
++ (void)captureEnvelope:(SentryEnvelope *)envelope;
+
++ (void)storeEnvelope:(SentryEnvelope *)envelope;
+
+@end
+
+NS_ASSUME_NONNULL_END